Fix: 소켓 로직 위치 수정
